+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the Halls of the Damned campaign website will be document
 
 Format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and this project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.7.3] - 2026-04-25
+## [2.8.0] - 2026-04-25
 
 ### Added
 - Added Session 27 Notes to /data